feat: fill balances field in dashboard group summary
See: BEDS-1065
