+++ backend/pkg/api/data_access/vdb_helpers.go
@@ -5,10 +5,12 @@ import (
 	"database/sql"
 	"fmt"
 	"math"
+	"math/big"
 	"time"
 
 	"github.com/doug-martin/goqu/v9"
 	"github.com/gobitfly/beaconchain/pkg/api/enums"
+	"github.com/gobitfly/beaconchain/pkg/api/services"
 	t "github.com/gobitfly/beaconchain/pkg/api/types"
 	"github.com/gobitfly/beaconchain/pkg/commons/cache"
 	"github.com/gobitfly/beaconchain/pkg/commons/utils"
@@ -285,4 +287,99 @@ func (d *DataAccessService) getElClAPR(ctx context.Context, dashboardId t.VDBId,
 	elIncome = elIncome.Mul(decimal.NewFromInt(1e18))
 
 	return elIncome, elAPR, clIncome, clAPR, nil
+}
+
+type RpOperatorInfo struct {
+	ValidatorIndex     uint64          `db:"validatorindex"`
+	NodeFee            float64         `db:"node_fee"`
+	NodeDepositBalance decimal.Decimal `db:"node_deposit_balance"`
+	UserDepositBalance decimal.Decimal `db:"user_deposit_balance"`
+}
+
+func (d *DataAccessService) getValidatorDashboardRpOperatorInfo(ctx context.Context, dashboardId t.VDBId) ([]RpOperatorInfo, error) {
+	var rpOperatorInfo []RpOperatorInfo
+
+	ds := goqu.Dialect("postgres").
+		Select(
+			goqu.L("v.validatorindex"),
+			goqu.L("rplm.node_fee"),
+			goqu.L("rplm.node_deposit_balance"),
+			goqu.L("rplm.user_deposit_balance")).
+		From(goqu.L("rocketpool_minipools AS rplm")).
+		LeftJoin(goqu.L("validators AS v"), goqu.On(goqu.L("rplm.pubkey = v.pubkey"))).
+		Where(goqu.L("node_deposit_balance IS NOT NULL")).
+		Where(goqu.L("user_deposit_balance IS NOT NULL"))
+
+	if len(dashboardId.Validators) == 0 {
+		ds = ds.
+			LeftJoin(goqu.L("users_val_dashboards_validators uvdv"), goqu.On(goqu.L("uvdv.validator_index = v.validatorindex"))).
+			Where(goqu.L("uvdv.dashboard_id = ?", dashboardId.Id))
+	} else {
+		ds = ds.
+			Where(goqu.L("v.validatorindex = ANY(?)", pq.Array(dashboardId.Validators)))
+	}
+
+	query, args, err := ds.Prepared(true).ToSQL()
+	if err != nil {
+		return nil, fmt.Errorf("error preparing query: %w", err)
+	}
+
+	err = d.alloyReader.SelectContext(ctx, &rpOperatorInfo, query, args...)
+	if err != nil {
+		return nil, fmt.Errorf("error retrieving rocketpool validators data: %w", err)
+	}
+	return rpOperatorInfo, nil
+}
+
+func (d *DataAccessService) calculateValidatorDashboardBalance(ctx context.Context, rpOperatorInfo []RpOperatorInfo, validators []t.VDBValidator, validatorMapping *services.ValidatorMapping, protocolModes t.VDBProtocolModes) (t.ValidatorBalances, error) {
+	balances := t.ValidatorBalances{}
+
+	rpValidators := make(map[uint64]RpOperatorInfo)
+	for _, res := range rpOperatorInfo {
+		rpValidators[res.ValidatorIndex] = res
+	}
+
+	// Create a new sub-dashboard to get the total cl deposits for non-rocketpool validators
+	var nonRpDashboardId t.VDBId
+
+	for _, validator := range validators {
+		metadata := validatorMapping.ValidatorMetadata[validator]
+		validatorBalance := utils.GWeiToWei(big.NewInt(int64(metadata.Balance)))
+		effectiveBalance := utils.GWeiToWei(big.NewInt(int64(metadata.EffectiveBalance)))
+
+		if rpValidator, ok := rpValidators[validator]; ok {
+			if protocolModes.RocketPool {
+				// Calculate the balance of the operator
+				fullDeposit := rpValidator.UserDepositBalance.Add(rpValidator.NodeDepositBalance)
+				operatorShare := rpValidator.NodeDepositBalance.Div(fullDeposit)
+				invOperatorShare := decimal.NewFromInt(1).Sub(operatorShare)
+
+				base := decimal.Min(decimal.Max(decimal.Zero, validatorBalance.Sub(rpValidator.UserDepositBalance)), rpValidator.NodeDepositBalance)
+				commission := decimal.Max(decimal.Zero, validatorBalance.Sub(fullDeposit).Mul(invOperatorShare).Mul(decimal.NewFromFloat(rpValidator.NodeFee)))
+				reward := decimal.Max(decimal.Zero, validatorBalance.Sub(fullDeposit).Mul(operatorShare).Add(commission))
+
+				operatorBalance := base.Add(reward)
+
+				balances.Total = balances.Total.Add(operatorBalance)
+			} else {
+				balances.Total = balances.Total.Add(validatorBalance)
+			}
+			balances.StakedEth = balances.StakedEth.Add(rpValidator.NodeDepositBalance)
+		} else {
+			balances.Total = balances.Total.Add(validatorBalance)
+
+			nonRpDashboardId.Validators = append(nonRpDashboardId.Validators, validator)
+		}
+		balances.Effective = balances.Effective.Add(effectiveBalance)
+	}
+
+	// Get the total cl deposits for non-rocketpool validators
+	if len(nonRpDashboardId.Validators) > 0 {
+		totalNonRpDeposits, err := d.GetValidatorDashboardTotalClDeposits(ctx, nonRpDashboardId)
+		if err != nil {
+			return balances, fmt.Errorf("error retrieving total cl deposits for non-rocketpool validators: %w", err)
+		}
+		balances.StakedEth = balances.StakedEth.Add(totalNonRpDeposits.TotalAmount)
+	}
+	return balances, nil
 }
